style(fe): 알러지 component 분리 (LUP-17)
